requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.11.5
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
pytest==7.4.3
httpx==0.25.2
beanie==1.23.6
scikit-learn>=1.2
pandas>=1.5
joblib
python-dateutil
numpy
bcrypt==3.2.2